+++ next-docs/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Change Log
 
+## 9.20.0
+
+### Minor Changes
+
+- Bump minor version
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.20.0
+  - @heathmont/moon-themes-tw@9.20.0
+  - @heathmont/moon-accordion@9.20.0
+  - @heathmont/moon-assets@9.20.0
+  - @heathmont/moon-charts@9.20.0
+  - @heathmont/moon-components@9.20.0
+  - @heathmont/moon-core@9.20.0
+  - @heathmont/moon-datepicker@9.20.0
+  - @heathmont/moon-draggabletable@9.20.0
+  - @heathmont/moon-icons@9.20.0
+  - @heathmont/moon-pagination@9.20.0
+  - @heathmont/moon-popover@9.20.0
+  - @heathmont/moon-select@9.20.0
+  - @heathmont/moon-sidebar@9.20.0
+  - @heathmont/moon-table@9.20.0
+  - @heathmont/moon-themes@9.20.0
+  - @heathmont/moon-utils@9.20.0
+  - @heathmont/moon-icons-tw@9.20.0
+  - @heathmont/moon-table-tw@9.20.0
+
 ## 9.19.0
 
 ### Minor Changes