+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 9.20.0
+
+### Minor Changes
+
+- Bump minor version
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.20.0
+
 ## 9.19.0
 
 ### Minor Changes